renamed to .ts and only handles the general case now
